+++ server/src/main/java/moment/comment/application/CommentService.java
@@ -4,10 +4,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import moment.comment.application.dto.request.CommentCreateRequest;
-import moment.comment.application.dto.response.CommentCreateResponse;
-import moment.comment.application.dto.response.MyCommentsResponse;
 import moment.comment.domain.Comment;
+import moment.comment.dto.request.CommentCreateRequest;
+import moment.comment.dto.response.CommentCreateResponse;
+import moment.comment.dto.response.MyCommentsResponse;
 import moment.comment.infrastructure.CommentRepository;
 import moment.global.exception.ErrorCode;
 import moment.global.exception.MomentException;
