+++ server/src/main/java/moment/comment/dto/response/EmojiDetailResponse.java
@@ -1,4 +1,4 @@
-package moment.comment.application.dto.response;
+package moment.comment.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import moment.reply.domain.Emoji;
