+++ server/src/main/java/moment/comment/presentation/CommentController.java
@@ -10,9 +10,9 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import moment.auth.presentation.AuthenticationPrincipal;
 import moment.comment.application.CommentService;
-import moment.comment.application.dto.request.CommentCreateRequest;
-import moment.comment.application.dto.response.CommentCreateResponse;
-import moment.comment.application.dto.response.MyCommentsResponse;
+import moment.comment.dto.request.CommentCreateRequest;
+import moment.comment.dto.response.CommentCreateResponse;
+import moment.comment.dto.response.MyCommentsResponse;
 import moment.global.dto.response.ErrorResponse;
 import moment.global.dto.response.SuccessResponse;
 import moment.user.dto.request.Authentication;
