+++ server/src/test/java/moment/comment/presentation/CommentControllerTest.java
@@ -8,10 +8,10 @@ import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import java.util.List;
 import moment.auth.infrastructure.JwtTokenManager;
-import moment.comment.application.dto.request.CommentCreateRequest;
-import moment.comment.application.dto.response.CommentCreateResponse;
-import moment.comment.application.dto.response.MyCommentsResponse;
 import moment.comment.domain.Comment;
+import moment.comment.dto.request.CommentCreateRequest;
+import moment.comment.dto.response.CommentCreateResponse;
+import moment.comment.dto.response.MyCommentsResponse;
 import moment.comment.infrastructure.CommentRepository;
 import moment.global.exception.ErrorCode;
 import moment.global.exception.MomentException;
